convert wei to eth

+++ src/main/kotlin/io/openfuture/api/domain/scaffold/ScaffoldSummaryDto.kt
@@ -1,5 +1,6 @@
 package io.openfuture.api.domain.scaffold
 
+import java.math.BigDecimal
 import java.math.BigInteger
 
 /**
@@ -9,7 +10,7 @@ data class ScaffoldSummaryDto(
         val description: String,
         val fiatAmount: String,
         val fiatCurrency: String,
-        val amount: BigInteger,
+        val amount: BigDecimal,
         val transactionIndex: BigInteger,
         val vendorAddress: String
 )
